hw10: невозможно перейти куда-либо без авторизации
